webapp/requirements.txt
Flask
Flask-WTF
Flask-JSON
pandas
spotipy
python-dotenv
requests